Move a couple functions from modutils -> core.utils
