+++ fooof/core/modutils.py
@@ -44,53 +44,6 @@ def safe_import(*args):
     return mod
 
 
-def get_obj_desc():
-    """Get dictionary specifying FOOOF object names and kind of attributes.
-
-    Returns
-    -------
-    attibutes : dict
-        Mapping of FOOOF object attributes, and what kind of data they are.
-    """
-
-    attributes = {'results' : ['background_params_', 'peak_params_', 'error_',
-                               'r_squared_', '_gaussian_params'],
-                  'settings' : ['peak_width_limits', 'max_n_peaks', 'min_peak_amplitude',
-                                'peak_threshold', 'background_mode'],
-                  'data' : ['power_spectrum', 'freq_range', 'freq_res'],
-                  'freq_info' : ['freq_range', 'freq_res'],
-                  'arrays' : ['freqs', 'power_spectrum', 'background_params_',
-                              'peak_params_', '_gaussian_params']}
-
-    return attributes
-
-
-def get_data_indices(background_mode):
-    """Get a dictionary mapping the column labels to indices in FOOOF data (FOOOFResults).
-
-    Parameters
-    ----------
-    background_mode : {'fixed', 'knee'}
-        Which approach taken to fit the background.
-
-    Returns
-    -------
-    indices : dict
-        Mapping for data columns to the column indices in which they appear.
-    """
-
-    indices = {
-        'CF'  : 0,
-        'Amp' : 1,
-        'BW'  : 2,
-        'intercept' : 0,
-        'knee'      : 1 if background_mode == 'knee' else None,
-        'slope'     : 1 if background_mode == 'fixed' else 2
-    }
-
-    return indices
-
-
 def docs_drop_param(docstring):
     """Drop the first parameter description for a string representation of a docstring.
